+++ MyKitchen/Utilities/EdamamApi/RecipeApi.swift
@@ -14,7 +14,7 @@ struct RecipeApi : Codable {
     let source : String?
     let url : String?
     let shareAs : String?
-    let yield : Int?
+    let yield : Double?
     let dietLabels : [String]?
     let healthLabels : [String]?
     let cautions : [String]?
